xorg-driver: imxfb_vivante: fix cursor destruction
This patch prevents calling xf86_cursors_fini when stopping X server if xrandr
is disabled. When xrandr is disable the vivante driver does not call
xf86CrtcConfigInit which initializes xf86CrtcConfigPrivateIndex. If
xf86_cursors_fini is called without xf86CrtcConfigPrivateIndex being
initialized it crashes with NULL pointer dereference of xf86_config so in this
way this is avoided.

Signed-off-by: Zahari Doychev <zahari.doychev@linux.com>