gst: fontconfig: fix fontconfig cflags
This patche fixs the cflags returned by pkg-config for the fontconfig library.
The previously returned flags are wrong and are causing the compilation of
the gstreameri-plugins-good to fail. The patch removes the @FREETYPE_CFLAGS@
dependency in the "fontconfig.pc.in" files. This is actually not necessary
because this flags are always added from the Requires field which in this case
contains the freetype2 library.

Signed-off-by: Zahari Doychev <zahari.doychev@linux.com>